Add user authentication
Add getUserInfo() to account.js, which returns the user's info from the database if the user is logged in.
This function is called in hooks.server.js so that we can easily check if the user is still logged in at each request.
If the user insn't logged in, we redirect him to the login page. The user will still have access to the register page.