feat(layout): add footer
Add a footer with 3 buttons to go to the page `/profile`, `/friends` and
`/parts/new`.

A page for the route `/parts/new` and for `/friends` has been created to
prove that the link go to these page. These page are empty because this
isn't the role of the feature (see #34 and #25).

Note: Some tests have been added. To make these tests realistics, the
tests have to simulate a logged user so it will take a valid uuid
session from the database. But unlike the non-tests js files, the tests
cannot understand import shortcut like `$lib` or `$env` so the usual
pool cannot be imported, so a temp pool is created to get the uuid
session.

To run the tests, the site has to run in backward and (in another
instance), to run the tests the command `npx playwright test` has to be
executed.

Fix: #37