fix(parts): issue with redirection
A non-connected use was redirected to `/` instead of `/login` when
trying to access the `/parts` route.

Also, the message to say that the user didn't do parts was before in
french, now in english.

The load function used a function that make a calls to the database,
instead of use the locals storage that has the data required.

Fix: #41