feat(profile): possbility to add/change avatar
Add a new input to give possibility to user to set/change it profile
picture.
The profile pictures are stored in `/static/profile-picture` with the
name `<user_id>.png`.

Note: Add the `/static/profile-picture/` folder in the gitignore to
prevent profile picture from development to be added in the repository.

Fix: #49