fix(test): short import unrecognized
When importing a file with short import like `$lib` or importing a file
that imported a file with a short import in the tests libraries, there
was an error that say that the module was unrecognized.

fix(test): short import unrecognized

When importing a file with short import like `$lib` or importing a file
that imported a file with a short import in the tests libraries, there
was an error that say that the module was unrecognized.