feature(layout): Add global css (#64)
* feature(layout): global css

Add a global style for all pages in app.css file.

The style foundation is defined globally for the site but there are still a few steps to complete this feature: define the css correctly on the other pages in relation to the global css and remove tailwind from the other pages and from the global css file.

Tests: Tests have been added. the test verify with a admin user the css properties in a admin page css-template (user is create dynamically and deleted after testing).
To test this feature we simulated a  admin user taking added temporarily to the database. A temporary pool is created to obtain the uuid.
Because the tests do not understand imports such as $lib or $env. It is also necessary to ensure that an .env file is present at the root of the project. Some functions for testing have been added.

Launching the tests: The site must be launched in the background and then install playwright npx playwright install and launch the tests with npx playwright test. 

Fix: #65